refactor: Move data loading hooks logic inside ChartLoadingWrapper
